Refactor renderer: Move cosmic background logic from env to renderer
